+++ package.json
@@ -10,7 +10,7 @@
     "test": "vitest --passWithNoTests"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "1.9.7",
+    "@reduxjs/toolkit": "2.2.1",
     "axios": "1.5.1",
     "classnames": "2.3.2",
     "history": "5.3.0",
@@ -19,7 +19,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-helmet-async": "1.3.0",
-    "react-redux": "8.1.3",
+    "react-redux": "9.1.0",
     "react-router-dom": "6.16.0"
   },
   "devDependencies": {
@@ -31,7 +31,7 @@
     "@types/leaflet": "1.9.3",
     "@types/react": "18.2.25",
     "@types/react-dom": "18.2.11",
-    "@types/react-redux": "7.1.27",
+    "@types/react-redux": "7.1.33",
     "@types/testing-library__jest-dom": "5.14.9",
     "@typescript-eslint/eslint-plugin": "6.7.4",
     "@typescript-eslint/parser": "6.7.4",
