+++ src/components/current-offers.tsx
@@ -22,7 +22,7 @@ export default function CurrentOffers({currentOffers, currentCity}: ListCardsPro
     <div className="cities__places-container container">
       <section className="cities__places places">
         <h2 className="visually-hidden">Places</h2>
-        <b className="places__found">{currentOffersNumber} places to stay in {currentCity}</b>
+        <b className="places__found">{currentOffersNumber} place{currentOffersNumber > 1 && 's'} to stay in {currentCity}</b>
         <form className="places__sorting" action="#" method="get">
           <span className="places__sorting-caption">Sort by</span>
           <span className="places__sorting-type" tabIndex={0}>
