--- src/store/action.ts
@@ -1,4 +1,0 @@
-import {createAction} from '@reduxjs/toolkit';
-import { TCityName } from '../const';
-
-export const setCity = createAction<TCityName>('offers/setCity');
