+++ src/store/index.ts
@@ -1,4 +1,9 @@
 import {configureStore} from '@reduxjs/toolkit';
-import {reducer} from './reducer';
+import { offersSlice } from './slices/offers';
 
-export const store = configureStore({reducer});
+
+export const store = configureStore({
+  reducer: {
+    [offersSlice.name]: offersSlice.reducer,
+  },
+});
